+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/apirequest/ApiRequestImpl.java
@@ -13,7 +13,7 @@ import static com.yahoo.bard.webservice.web.ErrorMessageFormat.INVALID_ASYNC_AFT
 import static com.yahoo.bard.webservice.web.ErrorMessageFormat.INVALID_INTERVAL_GRANULARITY;
 import static com.yahoo.bard.webservice.web.ErrorMessageFormat.INVALID_TIME_ZONE;
 import static com.yahoo.bard.webservice.web.ErrorMessageFormat.METRICS_NOT_IN_TABLE;
-import static com.yahoo.bard.webservice.web.ErrorMessageFormat.TABLE_GRANULARITY_MISMATCH;
+import static com.yahoo.bard.webservice.web.ErrorMessageFormat.TABLE_UNDEFINED;
 import static com.yahoo.bard.webservice.web.ErrorMessageFormat.TIME_ALIGNMENT;
 import static com.yahoo.bard.webservice.web.ErrorMessageFormat.UNKNOWN_GRANULARITY;
 
@@ -54,6 +54,8 @@ import org.joda.time.DateTimeZone;
 import org.joda.time.Duration;
 import org.joda.time.Interval;
 import org.joda.time.Period;
+import org.joda.time.PeriodType;
+import org.joda.time.ReadablePeriod;
 import org.joda.time.format.DateTimeFormatter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -656,17 +658,32 @@ public abstract class ApiRequestImpl implements ApiRequest {
             Granularity granularity,
             LogicalTableDictionary logicalTableDictionary
     ) throws BadApiRequestException {
-        LogicalTable generated = logicalTableDictionary.get(new TableIdentifier(tableName, granularity));
-
-        // check if requested logical table grain pair exists
-        if (generated == null) {
-            String msg = TABLE_GRANULARITY_MISMATCH.logFormat(granularity, tableName);
-            LOG.error(msg);
-            throw new BadApiRequestException(msg);
+        // No matching table name is found for this table
+        if (!logicalTableDictionary.getNames().contains(tableName)) {
+            LOG.debug(TABLE_UNDEFINED.logFormat(tableName));
+            throw new BadApiRequestException(TABLE_UNDEFINED.format(tableName));
         }
 
-        LOG.trace("Generated logical table: {} with granularity {}", generated, granularity);
-        return generated;
+        LogicalTable table = logicalTableDictionary.get(new TableIdentifier(tableName, granularity));
+        // No matching granularity is found for this table
+        if (table == null) {
+            List<ReadablePeriod> granularities = logicalTableDictionary.getGranularities(tableName);
+            String message = ErrorMessageFormat.GRANULARITY_NOT_SUPPORTED.format(
+                    granularity.getName(),
+                    tableName,
+                    logicalTableDictionary.getGranularities(tableName).isEmpty()
+                            ? "no granularities"
+                            : granularities.stream()
+                            .map(ReadablePeriod::getPeriodType)
+                            .map(PeriodType::getName)
+                            .collect(Collectors.joining(", "))
+            );
+            LOG.error(message);
+            throw new IllegalArgumentException(message);
+        }
+
+        LOG.trace("Generated logical table: {} with granularity {}", table, granularity);
+        return table;
     }
 
     @Override